refactor(api): move fapiRequest into lib/fapi.ts
Co-locate the FAPI passthrough request with the other FAPI helpers
(bootstrapDevBrowser, fetchUserSettings) in lib/fapi.ts, where it
already reached for decodePublishableKey and CLERK_JS_API_VERSION.

Promote the passthrough response shape to a shared `ApiResponse` type
in lib/fetch.ts so the moved function does not have to import upward
from commands/. commands/api/fapi.ts keeps the command-layer instance
and host resolution.

Addresses review feedback on #345.
