+++ packages/cli-core/src/commands/api/fapi.ts
@@ -1,24 +1,15 @@
 /**
- * Frontend API (FAPI) passthrough for `clerk api --fapi`.
+ * Instance + FAPI host resolution for `clerk api --fapi`.
  *
  * FAPI is the public API that clerk-js consumes. Its host is per-instance and
- * derived from the instance's publishable key, and the endpoints exposed here
- * (e.g. `/v1/environment`) are public, so no auth header is sent.
+ * derived from the instance's publishable key. The passthrough request itself
+ * lives in `lib/fapi.ts` (`fapiRequest`) alongside the other FAPI helpers.
  */
 
 import { resolveAppContext, resolveFetchedApplicationInstance } from "../../lib/config.ts";
-import { normalizeBapiPath } from "../../lib/bapi-command.ts";
-import {
-  CliError,
-  ERROR_CODE,
-  FapiError,
-  throwUsageError,
-  withApiContext,
-} from "../../lib/errors.ts";
-import { decodePublishableKey, CLERK_JS_API_VERSION } from "../../lib/fapi.ts";
-import { loggedFetch } from "../../lib/fetch.ts";
+import { CliError, ERROR_CODE, throwUsageError, withApiContext } from "../../lib/errors.ts";
+import { decodePublishableKey } from "../../lib/fapi.ts";
 import { fetchApplication, type ApplicationInstance } from "../../lib/plapi.ts";
-import type { BapiResponse } from "./bapi.ts";
 
 interface ResolveOptions {
   app?: string;
@@ -67,40 +58,4 @@ async function resolveInstance(options: ResolveOptions): Promise<ApplicationInst
 export async function resolveFapiHost(options: ResolveOptions): Promise<string> {
   const instance = await resolveInstance(options);
   return decodePublishableKey(instance.publishable_key).fapiHost;
-}
-
-export async function fapiRequest(options: {
-  method: string;
-  path: string;
-  fapiHost: string;
-  body?: string;
-}): Promise<BapiResponse> {
-  const url = new URL(`https://${options.fapiHost}${normalizeBapiPath(options.path)}`);
-  if (!url.searchParams.has("_clerk_js_version")) {
-    url.searchParams.set("_clerk_js_version", CLERK_JS_API_VERSION);
-  }
-
-  const headers: Record<string, string> = { Accept: "application/json" };
-  if (options.body) headers["Content-Type"] = "application/json";
-
-  const response = await loggedFetch(url, {
-    tag: "fapi",
-    method: options.method,
-    headers,
-    body: options.body,
-  });
-
-  if (!response.ok) {
-    throw await FapiError.fromResponse(response);
-  }
-
-  const rawBody = await response.text();
-  let body: unknown;
-  try {
-    body = JSON.parse(rawBody);
-  } catch {
-    body = rawBody;
-  }
-
-  return { status: response.status, headers: response.headers, body, rawBody };
 }
